okx-limit-bot/requirements.txt
flask
python-dotenv
okx-python-sdk-api